Updated Login
More elegant finding of username and password fields - no xpath;
Browser waits until login page is loaded - waiting max of 10 sec or until login section appears in browser - If somebody has problems this could be increased to max of 20 or even 30 sec;;
Browser waits after login to load my.noip.com page (this is not dynamic-dns page), but the page which loads after login - didn't had time to make it with explicit wait so I made it with implicit wait of 1 sec. For implicit wait some characteristic element should be used;

The code is tested and working by me, but would appreciate your test results for improvement.